Change KeyLockerDeviceMessage property type from int to long
Fixes #227

+++ CarWash.ClassLibrary/Models/KeyLockerDeviceMessage.cs
@@ -14,7 +14,7 @@ namespace CarWash.ClassLibrary.Models
         /// Gets or sets the integer value representing the state of all boxes in the key locker device.
         /// Each bit in this integer corresponds to the state of a specific box.
         /// </summary>
-        public int Inputs { get; set; }
+        public long Inputs { get; set; }
 
         /// <summary>
         /// Returns a list of boolean values representing the state of each box in the key locker.
